+++ webapp/routes/api/report.js
@@ -7,6 +7,4 @@ module.exports = function (app) {
   app.get(app.locals.BASE_URL + 'api/report/getAnalysisData', controller.getAnalysisData);
   app.get(app.locals.BASE_URL + 'api/report/getStaticData', controller.getStaticData);
   app.get(app.locals.BASE_URL + 'api/report/getDynamicData', controller.getDynamicData);
-  app.get(app.locals.BASE_URL + 'api/report/getAnalysisTotals', controller.getAnalysisTotals);
-  app.get(app.locals.BASE_URL + 'api/report/getDetailsAnalysisTotals', controller.getDetailsAnalysisTotals);
 };
